feat: make mutator folder system modular
