+++ src/lib/client/mutators/index.mutator.ts
@@ -1,0 +1,7 @@
+import { clientTodoMutators } from '@/lib/client/mutators/todo.mutator';
+
+import { M } from '@/models/mutator/index.model';
+
+export const clientMutators: M<'client'> = {
+  ...clientTodoMutators,
+};
